+++ src/PixelShader8.h
@@ -34,7 +34,7 @@ public:
 
 private:
 	// Update these values if required:
-	static const UINT NUMBER_FX = 26;
+	static const UINT NUMBER_FX = 27;
 	static const UINT SLIDERVALUE_COUNT = 7;
 	static const UINT FX_PARAM_COUNT = 5;
 	static const UINT NEWVERTICES_COUNT = 6;
@@ -212,6 +212,7 @@ private:
 		L"Flash",
 		L"Pattern1",
 		L"Pattern2",
+		L"SpatialDistorsion"
 	};
 };