+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.4.1 (2026-07-08)
 
+* Release on website
 * New pixels shader: 'SpatialDistorsion'
 * Some fixes for  'Pattern2'